Move `is_screen_black` from `_stbt/core.py` to `_stbt/black.py`
